Made loading the environment optional
loading the environment on the target machine is now set through the
load_env flag in runCommand() and runScript() methods.
the command input for runCommand now needs to be a list to allow for a
mix of named arguments and an undefined length of commands to run.
Added testLoadEnv() to the unittest set.